Fixes file names.

+++ enemy.cpp
@@ -1,9 +1,9 @@
-//Enemy.cpp
+// enemy.cpp
 //Auth: William Horn
 //4.10.2017
 //Definitions for Enemy class
 
-#include "Enemy.h"
+#include "enemy.h"
 
 #include <SFML/Graphics.hpp>
 using sf::Vector2;
@@ -12,7 +12,7 @@ using sf::Color;
 //Ctor
 Enemy::Enemy(Vector2<float> starting_pos,
              Vector2<float> starting_dir,
-             int health, int d) : EnemyShape(Vector2<float>(ENEMY_HEIGHT,ENEMY_WIDTH)),
+             int health, int d) : EnemyShape(Vector2<float>((float)ENEMY_HEIGHT,(float)ENEMY_WIDTH)),
                                   direction(starting_dir),
                                   hp(health),
                                   damage(d)
